fluent-fallback-cache: Fix waiting from multiple tasks.
The cache effectively multiplexes a stream, so we can't just rely on the
stream receiving end to wake up all the relevant tasks.

This fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1613705#c56 for
me.

Disclaimer: I'm not the most familiar with async rust, so take with a
grain of salt.