chore(build): Add rust-toolchain file
This file is useful for signalling which Rust toolchain this project
requires/expects. This is the legacy format file which works for old
(pre 1.23.0) versions of rustup. Given the current project MSRV (1.67.0
project wide) this is overkill and we *could* use rust-toolchain.toml
with an option for more meta data, but a) we don't need the extra data
at this point and b) there is quite a bit of CI tooling out there
including GH Actions that is much newer than 1.23.0 but still only
recognizes this legacy file format.