docs: Remove useless doc comment
This doc comment triggers the `clippy::empty_line_after_doc_comments`
lint. It could be converted to a normal comment, or the newline could be
deleted, but I don't think the comment provides any information not
already contained in the struct name, so there is no point in keeping
it.